feat: 멀티 스레드 적용 및 비동기 처리 로직 클래스 분리
