+++ src/main/java/codesquad/Main.java
@@ -1,11 +1,8 @@
 package codesquad;
 
-import codesquad.handler.HttpRequestHandler;
-
 public class Main {
     public static void main(String[] args) {
-        HttpRequestHandler httpRequestHandler = new HttpRequestHandler();
-        WebApplicationServer webApplicationServer = new WebApplicationServer(httpRequestHandler);
+        WebApplicationServer webApplicationServer = new WebApplicationServer();
         webApplicationServer.start();
     }
 }
