+++ src/main/java/codesquad/WebApplicationServer.java
@@ -1,36 +1,24 @@
 package codesquad;
 
 import codesquad.handler.HttpRequestHandler;
-import codesquad.http.HttpRequest;
-import codesquad.http.HttpRequestMapper;
-import codesquad.http.HttpResponse;
-import codesquad.http.HttpVersion;
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.OutputStream;
 import java.net.ServerSocket;
 import java.net.Socket;
-import java.util.Map;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class WebApplicationServer {
 
-    public static final String CRLF = "\r\n";
     public static final int PORT = 8080;
     private static final Logger logger = LoggerFactory.getLogger(WebApplicationServer.class);
 
-    private final HttpRequestHandler httpRequestHandler;
-
-    public WebApplicationServer(HttpRequestHandler httpRequestHandler) {
-        this.httpRequestHandler = httpRequestHandler;
-    }
+    private final ExecutorService executorService = Executors.newFixedThreadPool(10);
 
     public void start() {
         try (ServerSocket serverSocket = new ServerSocket(PORT)) {
-            System.out.println("Listening for connection on port 8080 ....");
+            logger.debug("Listening for connection on port 8080 ....");
             run(serverSocket);
         } catch (IOException e) {
             logger.error(e.getMessage(), e);
@@ -39,46 +27,14 @@ public class WebApplicationServer {
 
     private void run(ServerSocket serverSocket) {
         while (true) {
-            try (Socket socket = serverSocket.accept();
-                 InputStream inputStream = socket.getInputStream();
-                 OutputStream outputStream = socket.getOutputStream()
-            ) {
-                BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(inputStream));
-                HttpRequest httpRequest = HttpRequestMapper.from(bufferedReader);
-                HttpResponse httpResponse = httpRequestHandler.handle(httpRequest);
-                sendResponse(outputStream, httpResponse);
+            try {
+                Socket socket = serverSocket.accept();
+                HttpProcessor httpProcessor = new HttpProcessor(new HttpRequestHandler(), socket);
+                executorService.execute(httpProcessor::process);
             } catch (IOException | RuntimeException e) {
                 logger.error(e.getMessage(), e);
             }
         }
-    }
-
-    public void sendResponse(OutputStream outputStream, HttpResponse httpResponse) throws IOException {
-        sendStatus(outputStream, httpResponse);
-        sendHeaders(outputStream, httpResponse);
-        sendBody(outputStream, httpResponse);
-    }
-
-    private void sendStatus(OutputStream outputStream, HttpResponse httpResponse) throws IOException {
-        String status = HttpVersion.HTTP1_1.getName() + httpResponse.getHttpStatus().getCode()
-                + httpResponse.getHttpStatus().getRepresentation();
-        outputStream.write(status.getBytes());
-        outputStream.write(CRLF.getBytes());
-    }
-
-    private void sendHeaders(OutputStream outputStream, HttpResponse httpResponse) throws IOException {
-        Map<String, String> headers = httpResponse.getHeaders();
-        for (String headerName : headers.keySet()) {
-            String headerValue = headers.get(headerName);
-            String header = headerName + ": " + headerValue;
-            outputStream.write(header.getBytes());
-        }
-        outputStream.write(CRLF.getBytes());
-        outputStream.write(CRLF.getBytes());
-    }
-
-    private void sendBody(OutputStream outputStream, HttpResponse httpResponse) throws IOException {
-        outputStream.write(httpResponse.getBody().getBytes());
     }
 
 }
